Implemented dynamic UI routes to the start page
