+++ data/genres.json
@@ -2,13 +2,15 @@
   "data": [
     {
       "id": "1",
-      "genre": "Pop",
-      "data": "pop_songs"
+      "name": "Pop",
+      "data": "pop_songs",
+      "image": "pop-logo"
     },
     {
       "id": "2",
-      "genre": "Kpop",
-      "data": ""
+      "name": "Kpop",
+      "data": "kpop-logo",
+      "image": "kpop-logo"
     }
   ]
 }
